@@ -1,11 +1,11 @@
 from setuptools import setup, find_packages
 
-with open('README.md', 'r') as f:
+with open('README.md', 'r', encoding='utf-8') as f:
     long_description = f.read()
 
 dev_requirements = [
-    'mypy==0.812',
-    'pylint==2.6.2',
+    'mypy==0.942',
+    'pylint==2.13.4',
 ]
 
 docs_requirements = [
@@ -14,13 +14,13 @@ docs_requirements = [
 
 setup(
     name='nari',
-    version='0.0.1',
+    version='0.1.0',
     author='Nonowazu',
     author_email='oowazu.nonowazu@gmail.com',
     description='A small project to parse FFXIV network event data',
     long_description=long_description,
     long_description_content_type='text/markdown',
-    python_requires='>=3.8',
+    python_requires='>=3.10',
     packages=find_packages(),
     entry_points={
         'console_scripts': [
