@@ -1,53 +1,61 @@
 """Classes and functions for handling network logs from ACT"""
-from typing import Union
+from typing import Optional
 
 from nari.io.reader import Reader
-from nari.types.event import Type as EventType
-from nari.types.event.mappings import IdEventMapping
-from nari.types.event.base import Event
-from nari.util.exceptions import InvalidChecksum, EventNotFound
+from nari.io.reader.actlogutils import ID_MAPPINGS, ActEventType, date_from_act_timestamp, validate_checksum
+from nari.types.event import Event
+from nari.util.exceptions import EventNotFound
+from nari.io.reader.actlogutils.exceptions import InvalidActChecksum
+
 
 class ActLogReader(Reader):
     """Implementation of the Reader class for parsing ACT network logs"""
     def __init__(self, actlog_path, *, raise_on_checksum_failure=False, raise_on_invalid_id=False):
-        # might need a helper function, but eh. Also TODO: close file at some point to be nice
-        self.handle = open(actlog_path, 'r')
-        # Let's talk about this. ACT network logs seem to use two different indexes – one for
-        # network events, and one for other types of events. Without breaking the code too badly,
-        # might be good to identify if an ID is based off network or 'other' (memory events only?)
-        # and cycle through the correct indexes
+        # might need a helper function, but eh.
+        self.handle = open(actlog_path, 'r', encoding='utf-8') # pylint: disable=consider-using-with
         self.index = 1
         self.raise_on_checksum_failure = raise_on_checksum_failure
         self.raise_on_invalid_id = raise_on_invalid_id
 
-    def handle_line(self, line: str) -> Event:
-        """Handles an act-specific line"""
+    def __del__(self):
+        """Handles closing the file when the object undergoes garbage collection"""
+        self.handle.close()
+
+    def handle_line(self, line: str) -> Optional[Event]:
+        """Handles an ACT-specific line"""
         args = line.strip().split('|')
         id_ = int(args[0])
+
+        if self.raise_on_invalid_id and id_ not in ID_MAPPINGS:
+            raise EventNotFound(f"ACT id: {id_}")
+
+        if self.raise_on_checksum_failure:
+            if id_ in (ActEventType.memoryzonechange, ActEventType.version):
+                self.index = 1
+
+            if validate_checksum(line.strip(), self.index) is False:
+                raise InvalidActChecksum(f'Invalid checksum for line {line.strip()} with index {self.index})')
+
+            self.index += 1
+
         datestr = args[1]
+        timestamp = date_from_act_timestamp(datestr)
 
-        event = None
+        event: Optional[Event] = None
 
-        if id_ in IdEventMapping:
-            if id_ in [EventType.config.value]:
-                # special handling - this one doesn't have a checksum
-                event = IdEventMapping[id_](datestr, params=args[2:])
-            else:
-                event = IdEventMapping[id_](datestr, params=args[2:-1], index=self.index, checksum=args[-1])
-            if self.raise_on_checksum_failure and not event.valid_checksum():
-                raise InvalidChecksum(f'Checksum is invalid for event: {line} (index {self.index})')
-        elif self.raise_on_invalid_id:
-            raise EventNotFound(f'No event found for id {id}')
+        if id_ in ID_MAPPINGS:
+            event = ID_MAPPINGS[id_](timestamp, args[2:-1])
         else:
-            event = Event(datestr, params=args[2:-1], index=self.index, checksum=args[-1], id_=id_)
+            event = Event(timestamp)
 
-        self.index += 1
         return event
 
-
-    def read_next(self) -> Union[Event, None]:
-        """Returns an array of all the act log events from the file"""
-        line: str = self.handle.readline()
-        if len(line) == 0:
-            return None
-        return self.handle_line(line)
+    def read_next(self) -> Optional[Event]:
+        """Returns an array of all the ACT log events from the file"""
+        # keep reading until we get a non-None response from handle_line
+        while True:
+            line: str = self.handle.readline()
+            if len(line) == 0:
+                return None
+            if (event := self.handle_line(line)) is not None:
+                return event
