@@ -1,9 +1,9 @@
 """Collection of writing-related classes and utilities"""
 
 from abc import ABCMeta, abstractmethod
-from typing import List, Iterator, Optional
+from typing import Iterator
 
-from nari.types.event.base import Event
+from nari.types.event import Event
 
 
 class Writer(metaclass=ABCMeta):
