@@ -1,5 +1,6 @@
 """Base for the main types nari handles, including helper types for annotation purposes"""
 
-from typing import NewType
-
-HexStr = NewType('HexStr', str)
+HexStr = str
+"""A string that only has hexadecimal values"""
+Timestamp = int
+"""Number of milliseconds since the unix epoch"""
